+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5.2 (unreleased)
 
 - Implement workaround for issue related to interpreting dotted decimal AS number notation, [link](https://github.com/CiscoDevNet/terraform-provider-iosxr/issues/263)
+- Add `use_af_group` attribute to `iosxr_router_bgp_neighbor_group` resource and data source
 
 ## 0.5.1